修复ub

+++ src/x86_64/utils/atom_queue.c
@@ -17,27 +17,27 @@ atom_queue *create_atom_queue(uint64_t size) {
 
 bool atom_push(atom_queue *queue, uint8_t data) {
     if (queue == NULL) return false;
-    //ticket_lock(&queue->lock);
+    //spin_lock(queue->lock);
     uint64_t head = load(&queue->head);
     uint64_t next = (head + 1) & queue->mask;
     if (next == load(&queue->tail)) return false;
     *(&queue->buf[head]) = data;
     store(&queue->head, next);
     queue->length++;
-    // ticket_unlock(&queue->lock);
+    // spin_unlock(queue->lock);
     return true;
 }
 
-uint8_t atom_pop(atom_queue *queue) {
-    if (queue == NULL) return (uint8_t)-1;
-    if (queue->length <= 0) return (uint8_t)-1;
-    // ticket_lock(&queue->lock);
+int atom_pop(atom_queue *queue) {
+    if (queue == NULL) return -1;
+    if (queue->length <= 0) return -1;
+    // spin_lock(queue->lock);
     uint64_t tail = load(&queue->tail);
     if (tail == load(&queue->head)) return (uint8_t)-1;
     uint8_t data = queue->buf[tail];
     store(&queue->tail, (tail + 1) & queue->mask);
     queue->length--;
-    //ticket_unlock(&queue->lock);
+    //spin_unlock(queue->lock);
     return data;
 }
